# public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Management API",
    "version": "1.0.0",
    "description": "A comprehensive API for user authentication and management with role-based permissions",
    "contact": {
      "name": "API Support",
      "email": "app@carik.id"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
...
  },
  "paths": {
...
    "/v1/client": {
      "get": {
        "summary": "Get clients for authenticated user",
        "description": "Retrieve a list of clients. If user has 'canManage(\"clients\")' permission, shows all clients. Otherwise, shows only clients associated with the user through client_user_maps.",
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Search term to filter clients by name or description",
            "schema": {
              "type": "string",
              "example": "company"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clients retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Clients retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "clients": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "example": "01234567-89ab-cdef-0123-456789abcdef"
                              },
                              "name": {
                                "type": "string",
                                "example": "Company ABC"
                              },
                              "description": {
                                "type": "string",
                                "example": "Main client company"
                              },
                              "parentId": {
                                "type": "string",
                                "nullable": true,
                                "example": "01234567-89ab-cdef-0123-456789abcdef"
                              },
                              "metadata": {
                                "type": "object",
                                "nullable": true
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "statusId": {
                                "type": "integer",
                                "example": 1
                              },
                              "userRole": {
                                "type": "string",
                                "nullable": true,
                                "example": "admin"
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer",
                          "example": 5
                        },
                        "canManage": {
                          "type": "boolean",
                          "example": false,
                          "description": "Indicates if the user has manage permission for clients"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create new client",
        "description": "Create a new client (requires create permission on clients resource)",
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Company ABC"
                  },
                  "description": {
                    "type": "string",
                    "example": "Main client company"
                  },
                  "parentId": {
                    "type": "string",
                    "example": "01234567-89ab-cdef-0123-456789abcdef"
                  },
                  "metadata": {
                    "type": "object",
                    "example": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict - Client already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/client/{id}": {
      "get": {
        "summary": "Get specific client",
        "description": "Retrieve information about a specific client by ID (requires read permission on clients resource)",
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Client ID",
            "schema": {
              "type": "string",
              "example": "01234567-89ab-cdef-0123-456789abcdef"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update client",
        "description": "Update an existing client (requires edit permission on clients resource)",
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Client ID",
            "schema": {
              "type": "string",
              "example": "01234567-89ab-cdef-0123-456789abcdef"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Company Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description"
                  },
                  "parentId": {
                    "type": "string",
                    "example": "01234567-89ab-cdef-0123-456789abcdef"
                  },
                  "metadata": {
                    "type": "object",
                    "example": {}
                  },
                  "statusId": {
                    "type": "integer",
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found"
          },
          "409": {
            "description": "Conflict - Client name already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete client",
        "description": "Soft delete a client by setting status_id to 1 (requires manage permission on clients resource)",
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Client ID",
            "schema": {
              "type": "string",
              "example": "01234567-89ab-cdef-0123-456789abcdef"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}